fix: start sketching support for modernc sqlite driver
